api_to_database.py
the test data variable naming  is completely irrelevant you can just use the the main variable i was just doing it if it would work or not XD, this has been a small project of mine where getting inspiration from my current role where we take data from api. the code is shit but it works. will improve on it and have to study how to sync data if the database has stopped